Test security

See merge request !17

+++ src/main/java/com/tma/tt/api/jpa/CandidateJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Candidate;
 
-public interface CandidateJpaRepository extends CommonJpaRepository<Candidate, Integer> {
+public interface CandidateJpaRepository extends CommonJpaRepository<Candidate, Long> {
 }
