+++ src/main/java/com/tma/tt/api/jpa/ProjectJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Project;
 
-public interface ProjectJpaRepository extends CommonJpaRepository<Project, Integer> {
+public interface ProjectJpaRepository extends CommonJpaRepository<Project, Long> {
 }
