+++ src/main/java/com/tma/tt/api/jpa/QuestionJpaRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Question;
 
-public interface QuestionJpaRepository extends CommonJpaRepository<Question, Integer> {
+public interface QuestionJpaRepository extends CommonJpaRepository<Question, Long> {
 }
