+++ src/main/java/com/tma/tt/api/jpa/RequestJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Request;
 
-public interface RequestJpaRepository extends CommonJpaRepository<Request, Integer> {
+public interface RequestJpaRepository extends CommonJpaRepository<Request, Long> {
 }
