+++ src/main/java/com/tma/tt/api/jpa/SkillJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Skill;
 
-public interface SkillJpaRepository extends CommonJpaRepository<Skill, Integer> {
+public interface SkillJpaRepository extends CommonJpaRepository<Skill, Long> {
 }
