+++ src/main/java/com/tma/tt/api/jpa/StudyPlanJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.StudyPlan;
 
-public interface StudyPlanJpaRepository extends CommonJpaRepository<StudyPlan, Integer> {
+public interface StudyPlanJpaRepository extends CommonJpaRepository<StudyPlan, Long> {
 }
