+++ src/main/java/com/tma/tt/api/jpa/TaskDetailJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.TaskDetail;
 
-public interface TaskDetailJpaRepository extends CommonJpaRepository<TaskDetail, Integer> {
+public interface TaskDetailJpaRepository extends CommonJpaRepository<TaskDetail, Long> {
 }
