+++ src/main/java/com/tma/tt/api/jpa/TaskJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Task;
 
-public interface TaskJpaRepository extends CommonJpaRepository<Task, Integer> {
+public interface TaskJpaRepository extends CommonJpaRepository<Task, Long> {
 }
