+++ src/main/java/com/tma/tt/api/jpa/TaskTemplateJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.TaskTemplate;
 
-public interface TaskTemplateJpaRepository extends CommonJpaRepository<TaskTemplate, Integer> {
+public interface TaskTemplateJpaRepository extends CommonJpaRepository<TaskTemplate, Long> {
 }
