+++ src/main/java/com/tma/tt/api/jpa/UserFeedbackJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.UserFeedback;
 
-public interface UserFeedbackJpaRepository extends CommonJpaRepository<UserFeedback, Integer> {
+public interface UserFeedbackJpaRepository extends CommonJpaRepository<UserFeedback, Long> {
 }
