+++ src/main/java/com/tma/tt/api/jpa/UserTestJpaRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.UserTest;
 
-public interface UserTestJpaRepository extends CommonJpaRepository<UserTest, Integer> {
+public interface UserTestJpaRepository extends CommonJpaRepository<UserTest, Long> {
 }
