+++ src/main/java/com/tma/tt/api/jpa/UserTestResultJpaRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.UserTestResult;
 
-public interface UserTestResultJpaRepository extends CommonJpaRepository<UserTestResult, Integer> {
+public interface UserTestResultJpaRepository extends CommonJpaRepository<UserTestResult, Long> {
 }
