+++ src/main/java/com/tma/tt/api/repository/CsvCandidateRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.CsvCandidate;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface CsvCandidateRepository extends ResourceRepositoryV2<CsvCandidate, Integer> {
+public interface CsvCandidateRepository extends ResourceRepositoryV2<CsvCandidate, Long> {
 
 }
