+++ src/main/java/com/tma/tt/api/repository/DeliveryGroupRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.DeliveryGroup;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface DeliveryGroupRepository extends ResourceRepositoryV2<DeliveryGroup, Integer> {
+public interface DeliveryGroupRepository extends ResourceRepositoryV2<DeliveryGroup, Long> {
 
 }
