+++ src/main/java/com/tma/tt/api/repository/ProjectRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Project;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface ProjectRepository extends ResourceRepositoryV2<Project, Integer> {
+public interface ProjectRepository extends ResourceRepositoryV2<Project, Long> {
 
 }
