+++ src/main/java/com/tma/tt/api/repository/RequestRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Request;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface RequestRepository extends ResourceRepositoryV2<Request, Integer> {
+public interface RequestRepository extends ResourceRepositoryV2<Request, Long> {
 
 }
