+++ src/main/java/com/tma/tt/api/repository/RoleRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Role;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface RoleRepository extends ResourceRepositoryV2<Role, Integer> {
+public interface RoleRepository extends ResourceRepositoryV2<Role, Long> {
 
 }
