+++ src/main/java/com/tma/tt/api/repository/ScheduleRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Schedule;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface ScheduleRepository extends ResourceRepositoryV2<Schedule, Integer> {
+public interface ScheduleRepository extends ResourceRepositoryV2<Schedule, Long> {
 
 }
