+++ src/main/java/com/tma/tt/api/repository/StudyPlanRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.StudyPlan;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface StudyPlanRepository extends ResourceRepositoryV2<StudyPlan, Integer> {
+public interface StudyPlanRepository extends ResourceRepositoryV2<StudyPlan, Long> {
 
 }
