+++ src/main/java/com/tma/tt/api/repository/TaskAssignmentRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.TaskAssignment;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface TaskAssignmentRepository extends ResourceRepositoryV2<TaskAssignment, Integer> {
+public interface TaskAssignmentRepository extends ResourceRepositoryV2<TaskAssignment, Long> {
 
 }
