+++ src/main/java/com/tma/tt/api/repository/TaskDetailRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.TaskDetail;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface TaskDetailRepository extends ResourceRepositoryV2<TaskDetail, Integer> {
+public interface TaskDetailRepository extends ResourceRepositoryV2<TaskDetail, Long> {
 
 }
