+++ src/main/java/com/tma/tt/api/repository/TaskRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Task;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface TaskRepository extends ResourceRepositoryV2<Task, Integer> {
+public interface TaskRepository extends ResourceRepositoryV2<Task, Long> {
 
 }
