+++ src/main/java/com/tma/tt/api/repository/UserTestRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.UserTest;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface UserTestRepository extends ResourceRepositoryV2<UserTest, Integer> {
+public interface UserTestRepository extends ResourceRepositoryV2<UserTest, Long> {
 
 }
